Containerized TripleO Overcloud.
This change adds a containerized version of the overcloud for TripleO.
It's still a work in progress but I figured I'd put it up so people can
comment on how they want to see this advance.  It is in working order
here but may require some tweaks to properly integrate with the devtest
setup.  Please see the README-containers.md file for more information
on how to set this up.

This is using a RHEL Atomic image for now and requires that we
bootstrap the image with a container which contains the os-* programs.

Change-Id: I8bbe75abc3f5e160bd4371710cbac236ba8eebc9
Co-Authored-By: Ryan Hallisey <rhallisey@redhat.com>